ADDED PAYMENT FUNCTION
Updated the rev_order to create a receipt.txt (I don't know how to format  it properly. Please help me fix it. All the values needed are there)

payment function now works. it will ask for their credit card or g cash and you can input any amount you want. (To Shayne, I found it weird that this option to input the desired amount since it will normally just get the exact amount. PERO THE MORE COMPLICATED THE OUR PROGRAM IS, THE HIGHER ARE GRADE IS)